Logged all possible unexpected exceptions and corrected a KeyError
The KeyError was occurring on the scheduler.schedule function.
Particularly, on the generation of the minimum number of shifts per day
restrictions. This is because the doctors list should not be used, as it
is not granted that the associated doctor will have a shift
configuration.

Added some debug logs to keep track of restrictions generation.